docs(metadata): add metadata docs
