+++ packages/suite/docs/METADATA.md
@@ -1,0 +1,115 @@
+# Metadata (labeling)
+
+Metadata is a feature which allows user to associate persistent data with their accounts, transactions, addresses or even hidden wallets.
+Trezor Suite refers to metadata as to "labeling" in user interface. 
+
+## Data stores
+Because Trezor Suite is not a typical application with a backend server, data must be stored elsewhere. Currently supported providers are:
+- Dropbox
+- Google Drive
+
+Planned to be supported in future:
+- Local file system (desktop only)
+- SD card
+
+## Data structure in store 
+### version 1.0.0 (current)
+device metadata example
+```javascript
+{
+  "version": "1.0.0",
+  "walletLabel": "my hidden wallet label",
+},
+```
+account metadata example
+```javascript
+{
+  "version": "1.0.0",
+  "accountLabel": "my cool account label",
+  "outputLabels": {
+    "9f472739fa7034dfb9736fa4d98915f2e8ddf70a86ee5e0a9ac0634f8c1d0007": {
+      "0": "transaction 1"
+    },
+  },
+  "addressLabels": {
+    "bc1qannfxke2tfd4l7vhepehpvt05y83v3qsf6nfkk": "my cool address label",
+  }
+}
+```
+
+### version 2.0.0 (future)
+Each record will have timestamp which will allow user to resolve potential conflicts
+
+## Data encryption
+Data is stored in encrypted form using aes-256-gcm cipher.
+Master key for encryption is generated by users Trezor with constants defined in 
+[suite/src/actions/suite/constants/metadataConstants.ts](../src/actions/suite/constants/metadataConstants.ts#L18)
+Files encryption-decryption logic is located in 
+[suite/src/utils/suite/metadata.ts](../src/utils/suite/metadata.ts)
+
+## Where data lives in App
+Settings related data is defined in [suite/src/reducers/suite/metadataReducer](../src/reducers/suite/metadataReducer.ts) which contains
+```
+{
+ enabled: bool,
+ initiating: bool,
+ provider: {
+   type: "dropbox" | "google",
+   token: string,
+   user: string
+ }
+}
+``` 
+
+Metadata itself is divided into 2 groups (device metadata and account metadata). They are stored together with concrete records. 
+
+Device has metadata property:
+
+```
+{
+  /*
+   - disabled is initial state. user has not interacted with metadata before, metadata keys are not available
+   - enabled means that metadata is enabled for this device and application has metadata keys. App will try to download and decipher metadata
+   - cancelled means that user rejected cipherKeyValue call on device. 
+  */
+  status: "disabled" |"enabled" | "cancelled",
+  key: string,
+  fileName: string,
+  aesKey: string,
+  walletLabel: string
+}
+
+```
+
+Account has metadata property which is an object of following shape:
+
+```
+{
+  key: string(xpub),
+  fileName: string,
+  aesKey: string,
+  outputLabels: {
+    [string(txid)]: {
+      [number(outputIndex)]: string
+    },
+  },
+  addressLabels: {
+    [string(address)]: string,
+  },
+  accountLabel: string
+}
+```
+
+## User stories
+First time user:
+1. User opens App for the first time. Metadata is disabled. "Add label" buttons are present on mouse hover over labelable data.
+1. User clicks "Add label" button.
+1. Device metadata key is generated.
+1. Using device metadata key, account metadata keys are created.
+1. Open modal with metadata providers and connect.
+1. Fetch data from metadata provider and set interval for fetching data.
+1. Activate editable input.
+
+<!-- TODO:  -->
+<!-- Returning user: -->
+<!-- Disable metadata: -->
